+++ source/renderer/app/config/walletRestoreConfig.js
@@ -9,7 +9,7 @@ import type {
   ImportWalletStep,
 } from '../types/walletRestoreTypes';
 
-const { isIncentivizedTestnet } = global;
+const { isIncentivizedTestnet, isShelleyTestnet } = global;
 
 export const RESTORE_WALLET_STEPS: Array<RestoreWalletStep> = [
   'type',
@@ -23,16 +23,17 @@ export const IMPORT_WALLET_STEPS: EnumMap<string, ImportWalletStep> = {
   WALLET_SELECT_IMPORT: 'WalletSelectImport',
 };
 
-export const WALLET_KINDS: EnumMap<string, WalletKind> = isIncentivizedTestnet
-  ? {
-      DAEDALUS: 'Daedalus',
-      YOROI: 'Yoroi',
-      HARDWARE: 'Hardware',
-    }
-  : {
-      DAEDALUS: 'Daedalus',
-      YOROI: 'Yoroi',
-    };
+export const WALLET_KINDS: EnumMap<string, WalletKind> =
+  isIncentivizedTestnet && !isShelleyTestnet
+    ? {
+        DAEDALUS: 'Daedalus',
+        YOROI: 'Yoroi',
+        HARDWARE: 'Hardware',
+      }
+    : {
+        DAEDALUS: 'Daedalus',
+        YOROI: 'Yoroi',
+      };
 
 export const WALLET_DAEDALUS_KINDS: EnumMap<
   string,