+++ source/renderer/app/containers/splash/SplashNetworkPage.js
@@ -2,6 +2,7 @@
 import React, { Component } from 'react';
 import { observer, inject } from 'mobx-react';
 import type { InjectedProps } from '../../types/injectedPropsType';
+import SplashNetworkSTN from '../../components/splash/SplashNetworkSTN';
 import SplashNetworkITN from '../../components/splash/SplashNetworkITN';
 import SplashNetworkFlight from '../../components/splash/SplashNetworkFlight';
 
@@ -15,8 +16,20 @@ export default class SplashNetworkPage extends Component<Props> {
   render() {
     const { networkStatus: networkStatusActions } = this.props.actions;
     const { openExternalLink } = this.props.stores.app;
-    const { isIncentivizedTestnetTheme } = this.props.stores.profile;
-    const { isIncentivizedTestnet, isFlight } = global;
+    const {
+      isIncentivizedTestnetTheme,
+      currentLocale,
+    } = this.props.stores.profile;
+    const { isIncentivizedTestnet, isShelleyTestnet, isFlight } = global;
+    if (isShelleyTestnet) {
+      return (
+        <SplashNetworkSTN
+          onClose={() => networkStatusActions.toggleSplash.trigger()}
+          openExternalLink={openExternalLink}
+          currentLocale={currentLocale}
+        />
+      );
+    }
     if (isIncentivizedTestnet) {
       return (
         <SplashNetworkITN